blog-template: refactor server test utils
* move everything to server subfolder
* create separate mock and test utils
