--- server/enums/MongoError.ts
@@ -1,3 +1,0 @@
-export enum MongoError {
-  DuplicateKey = 11000,
-}
