+++ server/enums/MongoErrorCode.ts
@@ -1,0 +1,3 @@
+export enum MongoErrorCode {
+  DuplicateKey = 11000,
+}
