+++ test/server/auth/AuthController.test.ts
@@ -4,7 +4,7 @@ import { AuthController } from '@server/auth/AuthController';
 import { AuthService } from '@server/auth/AuthService';
 import { mockUpdatedUser } from '@test/server/user/mocks/mockUpdatedUser';
 import { mockUser } from '@test/server/user/mocks/mockUser';
-import { createMockReply } from '@test/utils/mocks/createMockReply';
+import { createMockReply } from '@test/server/mockUtils';
 
 const cookie = 'cookie';
 const userId = '1';
